pyproject.toml
[project]
name = "chess-pressure"
version = "0.1.2"
description = "Chess pressure heatmap visualization"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi",
    "uvicorn",
    "python-chess",
]

[project.scripts]
chess-pressure = "chess_pressure.app:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/chess_pressure"]
